* develop: (339 commits)
  Prevent traps from casting periodic spells. This can be removed until proper event handling.
  Move flying npcs to the ground on dead. Close #216
  Fixed height and liquid loading. This was implemented by me in b2ab9a8a31209b15fcd8946aebd383297637da34 and 62724471c04a458036993acb6d5aaa823140feef (sorry for that).
  Added output for LandHeight to .gps command
  Header include vector in CRitual
  Header includes
  Corrected include Whooops.... here we go
  Header includes and forward declaration in CThreads and Database
  Header includes in Storage and CallBack
  Header includes for ObjectStorage
  Header includes for CObjectFactory and GameObject.h
  Include QuestMgr.h in GameObject.h
  Header includes
  Headerinclude IUpdatable to Object.h
  Added set vehicle faction to the faction of the first player passenger Close #205
  Assign m_difficulty to uint32 to prevent errors. Close #221
  Added GameObject classes:  *GameObject_Door  *GameObject_Button  *GameObject_QuestGiver  *GameObject_Chest  *GameObject_Trap  *GameObject_SpellFocus  *GameObject_Goober  *GameObject_FishingNode  *GameObject_Ritual  *GameObject_SpellCaster  *GameObject_FishingHole  *GameObject_Destructible
  Added structure gor gotype trap and spell_focus
  Prevent attack with invalid faction. Added because gos with type GAMEOBJECT_TYPE_SPELL_FOCUS  casting spells of linked gos with type GAMEOBJECT_TYPE_TRAP instead of spawning them.
  Implemented creature_proto_difficulty column money
  ...

Conflicts:
	.gitignore
	src/world/Chat.h
	src/world/DBC/DBCStores.cpp
	src/world/DBC/DBCStores.h
	src/world/GameObject.cpp
	src/world/GameObject.h
	src/world/Level3.cpp
	src/world/Master.cpp
	src/world/Object.h
	src/world/ObjectMgr.cpp
	src/world/ObjectMgr.h
	src/world/ScriptMgr.h
	src/world/TransporterHandler.cpp

+++ src/scripts/EventScripts/Event_Hallows_End.cpp
@@ -1,6 +1,6 @@
 /*
  * AscEmu Framework based on ArcEmu MMORPG Server
- * Copyright (C) 2014-2015 AscEmu Team <http://www.ascemu.org>
+ * Copyright (C) 2014-2016 AscEmu Team <http://www.ascemu.org>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -129,7 +129,7 @@ class ShadeOfTheHorsemanAI : public MoonScriptCreatureAI
                 {
                     WPCount = 29;
                     WayPoints = WaypointGoldshire;
-                    for (int i = 0; i <= WPCount; ++i)
+                    for (uint8 i = 0; i <= WPCount; ++i)
                     {
                         AddWaypoint(CreateWaypoint(i, 0, WayPoints[i].addition, WayPoints[i]));
                     }
@@ -220,7 +220,7 @@ class HeadlessHorsemanWispInvisAI : public MoonScriptCreatureAI
         ParentClass::AIUpdate();
     }
 
-    MoonScriptCreatureAI*    mHeadlessHorseman;
+    MoonScriptCreatureAI* mHeadlessHorseman;
 };
 
 
@@ -228,7 +228,7 @@ class WaterBarrel : public GameObjectAIScript
 {
     public:
 
-        WaterBarrel(GameObject*  goinstance) : GameObjectAIScript(goinstance) {}
+        WaterBarrel(GameObject* goinstance) : GameObjectAIScript(goinstance) {}
         static GameObjectAIScript* Create(GameObject* GO) { return new WaterBarrel(GO); }
 
         void OnActivate(Player* pPlayer)
